* changed (int) to (size_t) in python files that cast pointers to int * point tracking in Plot2DWidget * ModuleTool and GraphicsView working better... needs testing

+++ trunk/c/API/Python/TC_Antimony_py.c
@@ -13,7 +13,7 @@ static PyObject * pytc_loadSBMLString(PyObject *self, PyObject *args)
 	tc_loadSBMLString(s);
 
 	Py_INCREF(Py_None);
-	return Py_None;	
+	return Py_None;
 }
 
 static PyObject * pytc_loadAntimonyString(PyObject *self, PyObject *args)
@@ -26,7 +26,7 @@ static PyObject * pytc_loadAntimonyString(PyObject *self, PyObject *args)
 	tc_loadAntimonyString(s);
 
 	Py_INCREF(Py_None);
-	return Py_None;	
+	return Py_None;
 }
 
 static PyObject * pytc_loadSBMLFile(PyObject *self, PyObject *args)
@@ -39,7 +39,7 @@ static PyObject * pytc_loadSBMLFile(PyObject *self, PyObject *args)
 	tc_loadSBMLFile(s);
 
 	Py_INCREF(Py_None);
-	return Py_None;	
+	return Py_None;
 }
 
 static PyObject * pytc_loadAntimonyFile(PyObject *self, PyObject *args)
@@ -52,7 +52,7 @@ static PyObject * pytc_loadAntimonyFile(PyObject *self, PyObject *args)
 	tc_loadAntimonyFile(s);
 
 	Py_INCREF(Py_None);
-	return Py_None;	
+	return Py_None;
 }
 
 static PyObject * pytc_getSBMLString(PyObject *self, PyObject *args)
@@ -68,7 +68,7 @@ static PyObject * pytc_getSBMLString(PyObject *self, PyObject *args)
 	{
 		return NULL;
 	}
-	
+
 	if (pylist)
 	{
 		if (PyList_Check(pylist) || PyTuple_Check(pylist))
@@ -76,27 +76,27 @@ static PyObject * pytc_getSBMLString(PyObject *self, PyObject *args)
 			isList = PyList_Check(pylist);
 			N = isList ? PyList_Size(pylist) : PyTuple_Size (pylist);
 		}
-		
+
 		if (N > 0)
 		{
 			array = malloc( (1+N) * sizeof(void*) );
 			array[N] = 0;
-			
-			for(i=0; i<N; ++i ) 
-			{ 
-				array[i] = isList ? (void*)((int)PyInt_AsLong( PyList_GetItem( pylist, i ) )) : (void*)((int)PyInt_AsLong( PyTuple_GetItem( pylist, i ) ));
+
+			for(i=0; i<N; ++i )
+			{
+				array[i] = isList ? (void*)((size_t)PyInt_AsLong( PyList_GetItem( pylist, i ) )) : (void*)((size_t)PyInt_AsLong( PyTuple_GetItem( pylist, i ) ));
 			}
 		}
 	}
-	
+
 	if (array == 0)
 	{
 		array = tc_allItems();
 	}
-	
+
 	s = tc_getSBMLString(array);
 	free(array);
-	
+
 	return Py_BuildValue("s",s);
 }
 
@@ -113,7 +113,7 @@ static PyObject * pytc_getAntimonyString(PyObject *self, PyObject *args)
 	{
 		return NULL;
 	}
-	
+
 	if (pylist)
 	{
 		if (PyList_Check(pylist) || PyTuple_Check(pylist))
@@ -121,27 +121,27 @@ static PyObject * pytc_getAntimonyString(PyObject *self, PyObject *args)
 			isList = PyList_Check(pylist);
 			N = isList ? PyList_Size(pylist) : PyTuple_Size (pylist);
 		}
-		
+
 		if (N > 0)
 		{
 			array = malloc( (1+N) * sizeof(void*) );
 			array[N] = 0;
-			
-			for(i=0; i<N; ++i ) 
-			{ 
-				array[i] = isList ? (void*)((int)PyInt_AsLong( PyList_GetItem( pylist, i ) )) : (void*)((int)PyInt_AsLong( PyTuple_GetItem( pylist, i ) ));
+
+			for(i=0; i<N; ++i )
+			{
+				array[i] = isList ? (void*)((size_t)PyInt_AsLong( PyList_GetItem( pylist, i ) )) : (void*)((size_t)PyInt_AsLong( PyTuple_GetItem( pylist, i ) ));
 			}
 		}
 	}
-	
+
 	if (array == 0)
 	{
 		array = tc_allItems();
 	}
-	
+
 	s = tc_getAntimonyString(array);
 	free(array);
-	
+
 	return Py_BuildValue("s",s);
 }
 
@@ -155,10 +155,10 @@ static PyObject * pytc_writeSBMLFile(PyObject *self, PyObject *args)
 		return NULL;
 
 	tc_writeSBMLFile(array,s);
-	
+
 	free(array);
 	Py_INCREF(Py_None);
-	return Py_None;	
+	return Py_None;
 }
 
 static PyObject * pytc_writeAntimonyFile(PyObject *self, PyObject *args)
@@ -173,6 +173,6 @@ static PyObject * pytc_writeAntimonyFile(PyObject *self, PyObject *args)
 
 	free(array);
 	Py_INCREF(Py_None);
-	return Py_None;	
+	return Py_None;
 }
