+++ trunk/c/API/Python/TC_AutoGeneRegulatoryTool_py.c
@@ -13,18 +13,18 @@ static PyObject * pytc_partsIn(PyObject *self, PyObject *args)
         return NULL;
 
 	array = tc_partsIn((void*)i);
-	
+
 	if (array)
 	{
 		len = 0;
-		
+
 		while (array[len] != 0) ++len;
-		
+
 		pylist = PyTuple_New(len);
-		
-		for (i=0; i<len; i++) 
+
+		for (i=0; i<len; i++)
 		{
-			item = Py_BuildValue("i",(int)(array[i]));
+			item = Py_BuildValue("i",(size_t)(array[i]));
 			PyTuple_SetItem(pylist, i, item);
 		}
 		free(array);
@@ -33,7 +33,7 @@ static PyObject * pytc_partsIn(PyObject *self, PyObject *args)
 	{
 		pylist = PyTuple_New(0);
 	}
-	
+
 	return pylist;
 }
 
@@ -46,20 +46,20 @@ static PyObject * pytc_partsUpstream(PyObject *self, PyObject *args)
 
 	if(!PyArg_ParseTuple(args, "i", &i))
         return NULL;
-	
+
 	array = tc_partsUpstream((void*)i);
-	
+
 	if (array)
 	{
 		len = 0;
-		
+
 		while (array[len] != 0) ++len;
-		
+
 		pylist = PyTuple_New(len);
-		
-		for (i=0; i<len; i++) 
+
+		for (i=0; i<len; i++)
 		{
-			item = Py_BuildValue("i",(int)(array[i]));
+			item = Py_BuildValue("i",(size_t)(array[i]));
 			PyTuple_SetItem(pylist, i, item);
 		}
 		free(array);
@@ -68,7 +68,7 @@ static PyObject * pytc_partsUpstream(PyObject *self, PyObject *args)
 	{
 		pylist = PyTuple_New(0);
 	}
-	
+
 	return pylist;
 }
 
@@ -80,21 +80,21 @@ static PyObject * pytc_partsDownstream(PyObject *self, PyObject *args)
 
 	if(!PyArg_ParseTuple(args, "i", &i))
         return NULL;
-		
-	
+
+
 	array = tc_partsDownstream((void*)i);
-	
+
 	if (array)
 	{
 		len = 0;
-		
+
 		while (array[len] != 0) ++len;
-		
+
 		pylist = PyTuple_New(len);
-		
-		for (i=0; i<len; i++) 
+
+		for (i=0; i<len; i++)
 		{
-			item = Py_BuildValue("i",(int)(array[i]));
+			item = Py_BuildValue("i",(size_t)(array[i]));
 			PyTuple_SetItem(pylist, i, item);
 		}
 		free(array);
@@ -103,6 +103,6 @@ static PyObject * pytc_partsDownstream(PyObject *self, PyObject *args)
 	{
 		pylist = PyTuple_New(0);
 	}
-	
+
 	return pylist;
 }
